requirements.txt
sphinx
Pillow
Flask
pytest
